pyproject.toml
[project]
name = "ai-detector"
version = "1.0.0"
description = "Unified AI-powered detection tool for identifying AI-generated images, videos, and deepfake audio."
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    # Shared
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "numpy>=2.2.4",
    "gradio>=5.23.1",

    # Image detection
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "accelerate>=1.5.2",
    "transformers>=4.50.2",
    "pillow-avif-plugin>=1.5.1",
    "pillow-heif>=0.22.0",
    "piexif>=1.1.3",
    "invisible-watermark>=0.2.0",

    # Audio detection
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "scipy>=1.11.4",
    "scikit-learn>=1.3.2",
    "joblib>=1.3.2",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
]

[project.scripts]
ai-detect = "ai_detector.cli:main"

[tool.uv]
upgrade = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "platform_system == 'Windows'" },
  { index = "pytorch-cu126", marker = "platform_system == 'Linux'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "platform_system == 'Windows'" },
  { index = "pytorch-cu126", marker = "platform_system == 'Linux'" },
]
torchaudio = [
  { index = "pytorch-cu126", marker = "platform_system == 'Windows'" },
  { index = "pytorch-cu126", marker = "platform_system == 'Linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --disable-warnings"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
